backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
fastapi-cli==0.0.4

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Redis
redis==5.0.4

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP - using nltk only (spacy conflicts with pydantic v2)
nltk==3.8.1

# AI / ML
transformers==4.40.1
sentencepiece==0.2.0
huggingface-hub==0.23.0

# TTS
gTTS==2.5.1

# HTTP Client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
